requirements.txt
amqp==2.5.2
asgiref==3.2.7
billiard==3.6.3.0
celery==4.4.2
Django==3.0.5
django-cleanup==4.0.0
django-crispy-forms==1.9.0
django-taggit==1.2.0
importlib-metadata==1.6.0
kombu==4.6.8
Pillow==7.1.1
pytz==2020.1
six==1.14.0
sqlparse==0.3.1
vine==1.3.0
whitenoise==5.0.1
zipp==3.1.0
